Safari-611.1.10/LayoutTests/inspector/debugger/evaluateOnCallFrame-emulateUserGesture-userIsInteracting-expected.txt

Tests for Debugger.evaluateOnCallFrame emulateUserGesture option.


== Running test suite: Debugger.evaluateOnCallFrame.emulateUserGesture.userIsInteracting
-- Running test case: EvaluateWithoutEmulatingUserGesture
Before debugger
User is NOT Interacting
After debugger
-- Running test case: EvaluateWithEmulatingUserGesture
Before debugger
User is Interacting
After debugger
